requirements.txt
asgiref==3.5.2
black==23.1.0
click==8.1.3
Django==4.1.4
django-environ==0.9.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-mptt-admin==2.5.1
djangorestframework==3.14.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
